Agrego parseo de hasChart
    refs:#373
