+++ src/components/exportable_card/charts/CardNullChart.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import { ICardChartProps } from '../FullCardChart';
+
+
+export default (props: ICardChartProps) =>
+    <div />
